// src/api/inventory/models/LocationCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hospital Equipment & Service Management API
 * API for managing hospital equipment inventory, service requests, and location tracking.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LocationCreate
 */
export interface LocationCreate {
    /**
     * 
     * @type {string}
     * @memberof LocationCreate
     */
    building: string;
    /**
     * 
     * @type {string}
     * @memberof LocationCreate
     */
    floor: string;
    /**
     * 
     * @type {string}
     * @memberof LocationCreate
     */
    department: string;
    /**
     * 
     * @type {string}
     * @memberof LocationCreate
     */
    room: string;
    /**
     * 
     * @type {string}
     * @memberof LocationCreate
     */
    description?: string;
}

/**
 * Check if a given object implements the LocationCreate interface.
 */
export function instanceOfLocationCreate(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "building" in value;
    isInstance = isInstance && "floor" in value;
    isInstance = isInstance && "department" in value;
    isInstance = isInstance && "room" in value;

    return isInstance;
}

export function LocationCreateFromJSON(json: any): LocationCreate {
    return LocationCreateFromJSONTyped(json, false);
}

export function LocationCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): LocationCreate {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'building': json['building'],
        'floor': json['floor'],
        'department': json['department'],
        'room': json['room'],
        'description': !exists(json, 'description') ? undefined : json['description'],
    };
}

export function LocationCreateToJSON(value?: LocationCreate | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'building': value.building,
        'floor': value.floor,
        'department': value.department,
        'room': value.room,
        'description': value.description,
    };
}