// src/api/inventory/models/LocationUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hospital Equipment & Service Management API
 * API for managing hospital equipment inventory, service requests, and location tracking.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LocationUpdate
 */
export interface LocationUpdate {
    /**
     * 
     * @type {string}
     * @memberof LocationUpdate
     */
    building: string;
    /**
     * 
     * @type {string}
     * @memberof LocationUpdate
     */
    floor: string;
    /**
     * 
     * @type {string}
     * @memberof LocationUpdate
     */
    department: string;
    /**
     * 
     * @type {string}
     * @memberof LocationUpdate
     */
    room: string;
    /**
     * 
     * @type {string}
     * @memberof LocationUpdate
     */
    description?: string;
}

/**
 * Check if a given object implements the LocationUpdate interface.
 */
export function instanceOfLocationUpdate(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "building" in value;
    isInstance = isInstance && "floor" in value;
    isInstance = isInstance && "department" in value;
    isInstance = isInstance && "room" in value;

    return isInstance;
}

export function LocationUpdateFromJSON(json: any): LocationUpdate {
    return LocationUpdateFromJSONTyped(json, false);
}

export function LocationUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): LocationUpdate {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'building': json['building'],
        'floor': json['floor'],
        'department': json['department'],
        'room': json['room'],
        'description': !exists(json, 'description') ? undefined : json['description'],
    };
}

export function LocationUpdateToJSON(value?: LocationUpdate | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'building': value.building,
        'floor': value.floor,
        'department': value.department,
        'room': value.room,
        'description': value.description,
    };
}